+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Audio/GenericMixerManager.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace ChoyUtilities {
+
+    public abstract class GenericMixerManager<T> : GenericSingleton<T> 
+    where T : MonoBehaviour{
+
+        // TODO
+
+    }
+
+}