+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Core/Generic/GenericSingleton.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 
 namespace ChoyUtilities {
 
+    [DisallowMultipleComponent]
     public abstract class GenericSingleton<T> : MonoBehaviour
         where T : MonoBehaviour {